+++ app/build.gradle
@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 21
-    buildToolsVersion "20.0.0"
+    compileSdkVersion 22
+    buildToolsVersion "22.0.1"
 
     defaultConfig {
-        applicationId "com.bigkoo.pickerviewdemo"
-        minSdkVersion 14
-        targetSdkVersion 21
+        applicationId "com.xuhuan.pickerviewdemo"
+        minSdkVersion 9
+        targetSdkVersion 22
         versionCode 1
         versionName "1.0"
     }
@@ -21,7 +21,7 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:21.0.3'
+    compile 'com.android.support:appcompat-v7:22.2.0'
     compile project(':pickerview')
-//    compile 'com.bigkoo:pickerview:1.0.2'
+//    compile 'com.xuhuan:pickerview:1.0.2'
 }
