+++ pickerview/build.gradle
@@ -2,17 +2,17 @@ apply plugin: 'com.android.library'
 apply plugin: 'com.github.dcendents.android-maven'
 apply plugin: 'com.jfrog.bintray'
 
-version = "1.0.3"
+version = "1.0.1"
 
 android {
-    compileSdkVersion 21
-    buildToolsVersion "20.0.0"
+    compileSdkVersion 22
+    buildToolsVersion "22.0.1"
 
     defaultConfig {
         minSdkVersion 9
-        targetSdkVersion 21
+        targetSdkVersion 22
         versionCode 1
-        versionName "1.0"
+        versionName version
     }
     buildTypes {
         release {
@@ -24,12 +24,12 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:21.0.3'
+    compile 'com.android.support:appcompat-v7:22.2.0'
 }
 
-def siteUrl = 'https://github.com/saiwu-bigkoo/Android-PickerView'                        // #CONFIG# // project homepage
-def gitUrl = 'https://github.com/saiwu-bigkoo/Android-PickerView.git'                     // #CONFIG# // project git
-group = "com.bigkoo"
+def siteUrl = 'https://github.com/xuhuan/Android-PickerView'                        // #CONFIG# // project homepage
+def gitUrl = 'https://github.com/xuhuan/Android-PickerView.git'                     // #CONFIG# // project git
+group = "com.xuhuan"
 
 install {
     repositories.mavenInstaller {
@@ -48,9 +48,10 @@ install {
                 }
                 developers {
                     developer {
-                        id 'sai'                                           // #CONFIG# // your user id (you can write your nickname)
-                        name 'sai.wu'                                       // #CONFIG# // your user name
-                        email 'sai.wu@bigkoo.com'                               // #CONFIG# // your email
+                        id 'xuhuan'
+                        // #CONFIG# // your user id (you can write your nickname)
+                        name 'xuhuan'                                       // #CONFIG# // your user name
+                        email 'xu.huan.web.master@gmail.com'                               // #CONFIG# // your email
                     }
                 }
                 scm {
@@ -81,6 +82,16 @@ task javadocJar(type: Jar, dependsOn: javadoc) {
 artifacts {
     archives javadocJar
     archives sourcesJar
+}
+
+javadoc {
+    options {
+        encoding "UTF-8"
+        charSet "UTF-8"
+        author true
+        version true
+        links "http://docs.oracle.com/javase/7/docs/api"
+    }
 }
 
 Properties properties = new Properties()
