+++ pickerview/src/androidTest/java/com/xuhuan/pickerview/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.bigkoo.pickerview;
+package com.xuhuan.pickerview;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
