+++ pickerview/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.bigkoo.pickerview">
+    package="com.xuhuan.pickerview">
 
 
 </manifest>
