+++ pickerview/src/main/res/layout/include_optionspicker.xml
@@ -1,25 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:background="@android:color/white"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal" >
-     <com.bigkoo.pickerview.lib.WheelView
-         android:id="@+id/options1"
-         android:layout_width="fill_parent"
-         android:layout_weight="1"
-         android:layout_height="wrap_content" />
+              android:layout_width="fill_parent"
+              android:layout_height="wrap_content"
+              android:background="@android:color/white"
+              android:orientation="horizontal">
 
-     <com.bigkoo.pickerview.lib.WheelView
-         android:id="@+id/options2"
-         android:layout_width="fill_parent"
-         android:layout_weight="1"
-         android:layout_height="wrap_content" />
+    <com.xuhuan.pickerview.lib.WheelView
+        android:id="@+id/options1"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"/>
 
-     <com.bigkoo.pickerview.lib.WheelView
-         android:id="@+id/options3"
-         android:layout_width="fill_parent"
-         android:layout_weight="1"
-         android:layout_height="wrap_content" />
-     
+    <com.xuhuan.pickerview.lib.WheelView
+        android:id="@+id/options2"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"/>
+
+    <com.xuhuan.pickerview.lib.WheelView
+        android:id="@+id/options3"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"/>
+
 </LinearLayout>
