+++ pickerview/src/main/res/layout/include_timepicker.xml
@@ -1,36 +1,37 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:background="@android:color/white"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal" >
-     <com.bigkoo.pickerview.lib.WheelView
-         android:id="@+id/year"
-         android:layout_width="wrap_content"
-         android:layout_weight="1"
-         android:layout_height="wrap_content" />
+              android:layout_width="fill_parent"
+              android:layout_height="wrap_content"
+              android:background="@android:color/white"
+              android:orientation="horizontal">
 
-     <com.bigkoo.pickerview.lib.WheelView
-         android:id="@+id/month"
-         android:layout_width="wrap_content"
-         android:layout_weight="1"
-         android:layout_height="wrap_content" />
+    <com.xuhuan.pickerview.lib.WheelView
+        android:id="@+id/year"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"/>
 
-     <com.bigkoo.pickerview.lib.WheelView
-         android:id="@+id/day"
-         android:layout_width="wrap_content"
-         android:layout_weight="1"
-         android:layout_height="wrap_content" />
-     
-     <com.bigkoo.pickerview.lib.WheelView
-         android:id="@+id/hour"
-         android:layout_width="wrap_content"
-         android:layout_weight="1"
-         android:layout_height="wrap_content" />
-     
-     <com.bigkoo.pickerview.lib.WheelView
-         android:id="@+id/min"
-         android:layout_width="wrap_content"
-         android:layout_weight="1"
-         android:layout_height="wrap_content" />
+    <com.xuhuan.pickerview.lib.WheelView
+        android:id="@+id/month"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"/>
+
+    <com.xuhuan.pickerview.lib.WheelView
+        android:id="@+id/day"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"/>
+
+    <com.xuhuan.pickerview.lib.WheelView
+        android:id="@+id/hour"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"/>
+
+    <com.xuhuan.pickerview.lib.WheelView
+        android:id="@+id/min"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"/>
 </LinearLayout>
